requirements.txt
amqp==2.5.0
anyjson==0.3.3
appdirs==1.4.3
asn1crypto==0.24.0
attrs==19.1.0
beautifulsoup4==4.6.0
billiard==3.6.0.0
black==19.3b0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cryptography==2.1.4
cupshelpers==1.0
Django==2.2.3
django-celery==3.3.0
django-redis==4.10.0
django-rest-framework==0.1.0
djangorestframework==3.9.4
entrypoints==0.3
Faker==1.0.7
flake8==3.7.7
html5lib==0.999999999
idna==2.8
keyring==12.2.1
keyrings.alt==3.0
kombu==4.6.3
lxml==4.3.4
mccabe==0.6.1
meld==3.18.0
onboard==1.4.1
pexpect==4.2.1
Pillow==5.1.0
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pyflakes==2.1.1
pygobject==3.28.2
pysmbc==1.0.15.6
python-apt==1.6.0
python-dateutil==2.8.0
python-xlib==0.20
pytz==2019.1
pyxdg==0.25
redis==3.2.1
reportlab==3.4.0
requests==2.22.0
SecretStorage==2.3.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
toml==0.10.0
unattended-upgrades==0.1
urllib3==1.25.3
vine==1.3.0
webencodings==0.5
